Fix trigger page inputs
